move button out of modal component
